requirements.txt
torch
torchvision
numpy
pillow
mss
pynput
